Fixed: DAO use in interactor, DAO use case, Saving, Interactor now uses DAO
