+++ src/main/java/use_case/BuildPokemonTeam/BuildPokemonTeamDataAccessInterface.java
@@ -2,6 +2,7 @@ package use_case.BuildPokemonTeam;
 import entity.Team;
 import entity.Pokemon;
 
+import java.io.FileNotFoundException;
 import java.util.Map;
 
 
@@ -17,6 +18,6 @@ public interface BuildPokemonTeamDataAccessInterface {
      * @param name name of the team
      * @return true if team exists
      */
-    boolean exists(String name);
+    boolean teamExists(Team name) throws FileNotFoundException;
 
 }
