+++ src/main/java/use_case/BuildPokemonTeam/BuildPokemonTeamInputBoundary.java
@@ -15,4 +15,6 @@ public interface BuildPokemonTeamInputBoundary {
     void addToTeam(BuildPokemonTeamInputData BuildPokemonTeamInputData) throws IOException;
 
     void removeFromTeam(BuildPokemonTeamInputData BuildPokemonTeamInputData) throws IOException;
+
+    void saveTeam(BuildPokemonTeamInputData BuildPokemonTeamInputData) throws IOException;
 }
